# Shorter literal/match length reading in amd64 decoder