This mainly makes the assembly code easier to read, but it also skips two memory accesses per token. Performance is practically unchanged (Go 1.17-rc1):

```
name                old time/op    new time/op    delta
Uncompress-8          8.76ns ± 1%    8.76ns ± 1%    ~     (p=0.870 n=18+18)
UncompressPg1661-8     359µs ± 1%     358µs ± 1%  -0.38%  (p=0.035 n=20+20)
UncompressDigits-8    51.8µs ± 1%    51.8µs ± 2%    ~     (p=0.652 n=20+19)
UncompressTwain-8      236µs ± 1%     234µs ± 1%  -0.85%  (p=0.000 n=19+19)
UncompressRand-8      4.22µs ± 2%    4.22µs ± 2%    ~     (p=0.909 n=20+20)

name                old alloc/op   new alloc/op   delta
Uncompress-8           0.00B          0.00B         ~     (all equal)
UncompressPg1661-8      184B ± 0%      184B ± 0%    ~     (all equal)
UncompressDigits-8      190B ± 9%      186B ± 5%    ~     (p=0.131 n=20+19)
UncompressTwain-8       184B ± 0%      184B ± 0%    ~     (all equal)
UncompressRand-8        186B ± 1%      186B ± 1%    ~     (p=0.080 n=19+20)

name                old allocs/op  new allocs/op  delta
Uncompress-8            0.00           0.00         ~     (all equal)
UncompressPg1661-8      4.00 ± 0%      4.00 ± 0%    ~     (all equal)
UncompressDigits-8      4.00 ± 0%      4.00 ± 0%    ~     (all equal)
UncompressTwain-8       4.00 ± 0%      4.00 ± 0%    ~     (all equal)
UncompressRand-8        4.00 ± 0%      4.00 ± 0%    ~     (all equal)

name                old speed      new speed      delta
UncompressPg1661-8  1.05GB/s ± 1%  1.05GB/s ± 1%  +0.38%  (p=0.033 n=20+20)
UncompressDigits-8  1.84GB/s ± 1%  1.84GB/s ± 2%    ~     (p=0.647 n=20+19)
UncompressTwain-8   1.09GB/s ± 1%  1.10GB/s ± 1%  +0.86%  (p=0.000 n=19+19)
UncompressRand-8    3.89GB/s ± 2%  3.88GB/s ± 2%    ~     (p=0.925 n=20+20)
```